moving temp log to util
